This pull request migrates the code to use `xp-framework/reflection`. As this changes the API, we will releases this as a major release, 5.0.0 at the time of writing.

## Performance

No measurable impact.

### Before

```bash
$ xp test src/test/php
# ...
Test cases:  156 succeeded
Memory used: 3847.27 kB (3901.40 kB peak)
Time taken:  0.078 seconds (0.159 seconds overall)
```

### After

```bash
$ xp test src/test/php
# ...
Test cases:  156 succeeded
Memory used: 3639.90 kB (3694.02 kB peak)
Time taken:  0.079 seconds (0.157 seconds overall)
```

## See also

* https://github.com/xp-framework/rfc/issues/338
* https://github.com/xp-framework/core/pull/332