@thekid 2023-07-22
Released in https://github.com/xp-forge/frontend/releases/tag/v5.0.0